stack-to-nix: Split into modules
Roughly dividing up the code into modules. No attempts are made to
clean anything up.
